requirements.txt
langchain_openai
langchain_core
langchain_groq
python-dotenv
streamlit
langchain_community
langserve
fastapi
uvicorn
requests
bs4
pypdf
chromadb
faiss-cpu
arxiv
wikipedia
langchainhub
langchain-objectbox